+++ src/fastq.hpp
@@ -111,10 +111,12 @@ public:
      *
      * @param trim_ns If true, ambiguous bases ('N') are trimmed.
      * @param low_quality Trim bases with a quality score at or below this value.
+     * @param preserve5p Only trim from the 3p end if true.
      * @return A pair containing the number of 5' and 3' bases trimmed.
      */
     ntrimmed trim_trailing_bases(const bool trim_ns = true,
-                                    char low_quality = -1);
+                                 char low_quality = -1,
+                                 const bool preserve5p = false);
 
     /**
      * Trims low-quality bases using a sliding window approach.
@@ -122,11 +124,13 @@ public:
      * @param trim_ns If true, ambiguous bases ('N') are trimmed.
      * @param low_quality Trim bases with a quality score at or below this value.
      * @param winlen The length of the sliding window.
+     * @param preserve5p Only trim from the 3p end if true.
      * @return A pair containing the number of 5' and 3' bases trimmed.
      */
     ntrimmed trim_windowed_bases(const bool trim_ns = true,
                                  char low_quality = -1,
-                                 const double window_size = 0.1);
+                                 const double window_size = 0.1,
+                                 const bool preserve5p = false);
 
 
     /**